connected our form to firebase and displayed firebase data in leaderboard
