+++ src/Game.js
@@ -104,7 +104,7 @@ const Game = () => {
       </div>
 
       {rounds >= 10 ? (
-        <Form />
+        <Form score={score} />
       ) : (
         showDefinition && (
           <div>
